Patch 0092
・アイテムトグルが固定化されるように修正
	取得したアイテムはインベントリの最後尾に格納される必要がありますが、現在の仕様では配列内で空いている場所
	に格納されるようになっています。
	従って clif_parse_LoadEndAck() でアイテムリストを再送信したときに以前と並びが異なるアイテムリストが
	クライアントに渡されてしまい、トグルが固定化されない原因となっていました。
	そこでアイテムを取得した順に番号を付け、pc_checkitem() でこの番号順にソートするように変更しました。
	item構造体のidメンバがほとんど使われてなかったので、この番号を保存する変数として利用しています。
	pc_checkitem() で行っていたアイテムの空き詰め処理の代わりにqsortを行う形になっていますが、インベントリに
	関してはソート後にアイテムデータベースへのポインタを再設定しないといけないので、以前に比べて若干処理は
	増えていることになります。
	-> item構造体の id メンバを unsigned int型に変更
	-> map_session_data構造体内、state構造体に inventory_dirty, cart_dirty 追加
	   （qsortすべきかどうかを判定するフラグ）
	-> map_session_data構造体に inventory_sortkey, cart_sortkey 追加
	-> pc_setinventorydata() の処理を pc_checkitem() 内に吸収
・倉庫内アイテムの並び換えを修正および拡張
	上記のアイテムトグルと同様に倉庫内アイテムについても格納順に並ぶようにソートするように修正しました。
	battle_athena.confでソート方法を選択することができます。
	ただし今までの storagesort_by_itemid は削除されて personal_storage_sort, guild_storage_sort という項目に
	変更されているので注意してください。
	0のときはソートしません。速度重視。
	1のときはitem構造体のidを使って格納順にソートします。本鯖準拠。
	2のときはアイテムIDでソートします。091以前のバージョンの方法です。
	3のときは種類でソートします。まずtypeでソートし、次に武器・防具ならLoc（装備箇所）でソートし、
	最後にアイテムIDでソートします。これにより同じタイプのアイテムが近くに集まるようになります。
	-> storage構造体、guild_storage構造体に sortkey 追加
	-> storage_comp_item_by_key(), storage_comp_item_by_nameid(), storage_comp_item_by_type() 追加
	-> sortage_gsortitem() を削除し storage_sortitem() で一括処理できるように変更
・SQL: sql-filesにAuriga092_changetable.sql追加、main.sql更新
       `id` フィールドの使い方が変更されているので注意してください
・SQL: item構造体のidメンバが正しくコンバートされるように修正（char-converter.c）

+++ src/char/int_storage.c
@@ -79,7 +79,7 @@ int storage_tostr(char *str,struct storage *p)
 
 	for(i=0;i<MAX_STORAGE;i++)
 		if( (p->store_item[i].nameid) && (p->store_item[i].amount) ){
-			str_p += sprintf(str_p,"%d,%d,%d,%d,%d,%d,%d,%d,%d,%d,%d ",
+			str_p += sprintf(str_p,"%u,%d,%d,%d,%d,%d,%d,%d,%d,%d,%d ",
 				p->store_item[i].id,p->store_item[i].nameid,p->store_item[i].amount,p->store_item[i].equip,
 				p->store_item[i].identify,p->store_item[i].refine,p->store_item[i].attribute,
 				p->store_item[i].card[0],p->store_item[i].card[1],p->store_item[i].card[2],p->store_item[i].card[3]);
@@ -106,17 +106,17 @@ int storage_fromstr(char *str,struct storage *p)
 	if(set!=2)
 		return 1;
 	if(str[next]=='\n' || str[next]=='\r')
-		return 0;	
+		return 0;
 	next++;
 	for(i=0;str[next] && str[next]!='\t';i++){
-		set=sscanf(str+next,"%d,%d,%d,%d,%d,%d,%d,%d,%d,%d,%d%n",
+		set=sscanf(str+next,"%u,%d,%d,%d,%d,%d,%d,%d,%d,%d,%d%n",
 		  &tmp_int[0],&tmp_int[1],&tmp_int[2],&tmp_int[3],
 		  &tmp_int[4],&tmp_int[5],&tmp_int[6],
 		  &tmp_int[7],&tmp_int[8],&tmp_int[9],&tmp_int[10],&len);
 		if(set!=11)
 			return 1;
 		if(i < MAX_STORAGE) {
-			p->store_item[i].id=tmp_int[0];
+			p->store_item[i].id=(unsigned int)tmp_int[0];
 			p->store_item[i].nameid=tmp_int[1];
 			p->store_item[i].amount=tmp_int[2];
 			p->store_item[i].equip=tmp_int[3];
@@ -179,10 +179,10 @@ int storage_txt_sync(void)
 {
 	FILE *fp;
 	int lock;
-	
+
 	if( !storage_db )
 		return 1;
-	
+
 	if( (fp=lock_fopen(storage_txt,&lock))==NULL ){
 		printf("int_storage: cant write [%s] !!! data is lost !!!\n",storage_txt);
 		return 1;
@@ -234,7 +234,7 @@ static int gstorage_tostr(char *str,struct guild_storage *p)
 
 	for(i=0;i<MAX_GUILD_STORAGE;i++)
 		if( (p->store_item[i].nameid) && (p->store_item[i].amount) ){
-			str_p += sprintf(str_p,"%d,%d,%d,%d,%d,%d,%d,%d,%d,%d,%d ",
+			str_p += sprintf(str_p,"%u,%d,%d,%d,%d,%d,%d,%d,%d,%d,%d ",
 				p->store_item[i].id,p->store_item[i].nameid,p->store_item[i].amount,p->store_item[i].equip,
 				p->store_item[i].identify,p->store_item[i].refine,p->store_item[i].attribute,
 				p->store_item[i].card[0],p->store_item[i].card[1],p->store_item[i].card[2],p->store_item[i].card[3]);
@@ -260,17 +260,17 @@ static int gstorage_fromstr(char *str,struct guild_storage *p)
 	if(set!=2)
 		return 1;
 	if(str[next]=='\n' || str[next]=='\r')
-		return 0;	
+		return 0;
 	next++;
 	for(i=0;str[next] && str[next]!='\t';i++){
-		set=sscanf(str+next,"%d,%d,%d,%d,%d,%d,%d,%d,%d,%d,%d%n",
+		set=sscanf(str+next,"%u,%d,%d,%d,%d,%d,%d,%d,%d,%d,%d%n",
 		  &tmp_int[0],&tmp_int[1],&tmp_int[2],&tmp_int[3],
 		  &tmp_int[4],&tmp_int[5],&tmp_int[6],
 		  &tmp_int[7],&tmp_int[8],&tmp_int[9],&tmp_int[10],&len);
 		if(set!=11)
 			return 1;
 		if(i < MAX_GUILD_STORAGE) {
-			p->store_item[i].id=tmp_int[0];
+			p->store_item[i].id=(unsigned int)tmp_int[0];
 			p->store_item[i].nameid=tmp_int[1];
 			p->store_item[i].amount=tmp_int[2];
 			p->store_item[i].equip=tmp_int[3];
@@ -597,7 +597,7 @@ void gstorage_txt_final(void)
 {
 	if(gstorage_db)
 		numdb_final(gstorage_db,gstorage_db_final);
-		
+
 #ifdef TXT_JOURNAL
 	if( guild_storage_journal_enable )
 	{
